# backend/docs/swagger.yaml
basePath: /
definitions:
...
  energy.AnalyticsPoint:
    properties:
      label:
        type: string
      load:
        type: number
      solar:
        type: number
    type: object
  energy.EnergyAnalytics:
    properties:
      battery_level:
        $ref: '#/definitions/energy.EnergyStats'
      consumed:
        $ref: '#/definitions/energy.EnergyStats'
      generated:
        $ref: '#/definitions/energy.EnergyStats'
      grid_export:
        $ref: '#/definitions/energy.EnergyStats'
      savings:
        $ref: '#/definitions/energy.SavingsStats'
    type: object
  energy.EnergyStats:
    properties:
      average:
        type: number
      max:
        type: number
      min:
        type: number
      total:
        type: number
    type: object
  energy.SavingsStats:
    properties:
      co2_reduction_kg:
        type: number
      estimated_savings_rs:
        type: number
      total_energy_kwh:
        type: number
    type: object
  govt.UpdateSubsidyRequest:
    properties:
      reason:
        description: Optional reason for rejection
        type: string
      status:
        enum:
        - APPROVED
        - REJECTED
        type: string
    required:
    - status
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API documentation for the Solar Energy Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Solar Energy Management System API
  version: "1.0"
paths:
...
  /govt/dashboard/stats:
    get:
      consumes:
      - application/json
      description: Get statistics for the government dashboard (counts of pending,
        approved, rejected, disbursed)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
      security:
      - BearerAuth: []
      summary: Get government dashboard stats
      tags:
      - Government
  /govt/subsidies/{id}/status:
    put:
      consumes:
      - application/json
      description: Approve or Reject a subsidy application
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/govt.UpdateSubsidyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update subsidy status
      tags:
      - Government
  /govt/subsidies/history:
    get:
      consumes:
      - application/json
      description: Get a history of processed subsidies (Approved, Rejected, Disbursed)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get subsidy history
      tags:
      - Government
  /govt/subsidies/pending:
    get:
      consumes:
      - application/json
      description: Get a list of users with PENDING subsidy status
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get pending subsidies
      tags:
      - Government
...
  /superadmin/energy/analytics:
    get:
      consumes:
      - application/json
      description: Get energy analytics and statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/energy.EnergyAnalytics'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get energy analytics
      tags:
      - Energy
  /superadmin/energy/history:
    get:
      consumes:
      - application/json
      description: Get historical energy data
      parameters:
      - description: Number of days (default 7)
        in: query
        name: days
        type: integer
      - description: Device ID
        in: query
        name: device_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get energy history
      tags:
      - Energy
  /superadmin/energy/trend:
    get:
      consumes:
      - application/json
      description: Get aggregated energy trends (Super Admin only)
      parameters:
      - description: Region filter
        in: query
        name: region
        type: string
      - description: Time period (week, month, quarter, year)
        in: query
        name: period
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              items:
                $ref: '#/definitions/energy.AnalyticsPoint'
              type: array
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get global energy trend
      tags:
      - Energy
...
  /user/energy/analytics:
    get:
      consumes:
      - application/json
      description: Get energy analytics and statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              $ref: '#/definitions/energy.EnergyAnalytics'
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get energy analytics
      tags:
      - Energy
  /user/energy/history:
    get:
      consumes:
      - application/json
      description: Get historical energy data
      parameters:
      - description: Number of days (default 7)
        in: query
        name: days
        type: integer
      - description: Device ID
        in: query
        name: device_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get energy history
      tags:
      - Energy
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"